restructured the dashboard layout
