+++ src/App.jsx
@@ -20,6 +20,7 @@ import Settings from './components/settings/Settings.jsx';
 import Chatbot from './components/chatbot/Chatbot.jsx';
 import Map from './components/map/Map.jsx';
 import TrendingArtists from './components/trendingartists/TrendingArtists.jsx';
+import DashboardLayout from './components/dashboardlayout/DashboardLayout.jsx';
 
 function App() {
     const [JWT, setJWT] = useState(null);
@@ -48,56 +49,35 @@ function App() {
         fetchUserInfo();
     }, [JWT, navigate]);
 
+    const handleLogout = () => {
+        localStorage.removeItem('jwt');
+        localStorage.removeItem('spotifyAuth');
+        setJWT(null);
+        navigate('/login');
+    };
+
     return (
         <>
             <Routes>
                 <Route path="/" element={<Landing />} />
                 <Route path='/login' element={<Login setJWT={setJWT} />} />
                 <Route path='/map' element={<Map />} />
-                <Route path="/profile" element={
-                    <RequireAuth>
-                        <Profile userInfo={userInfo} />
-                    </RequireAuth> } />
-                <Route path="/friends" element={
-                    <RequireAuth>
-                        <Friends userInfo={userInfo} />
-                    </RequireAuth> } />
-                <Route path="/trending" element={
-                    <RequireAuth>
-                        <TrendingArtists userInfo={userInfo} />
-                    </RequireAuth> } />
-                <Route path="/favourites" element={
-                    <RequireAuth>
-                        <Favourites userInfo={userInfo} />
-                    </RequireAuth> } />
-                <Route path="/l/dashboard" element={
-                    <RequireAuth>
-                        <ListenerDashboard userInfo={userInfo} />
-                    </RequireAuth> } />
-                <Route path="/tags" element={
-                    <RequireAuth>
-                        <TaggingScreen userInfo={userInfo} />
-                    </RequireAuth> } />
-                <Route path="/chatbot" element={
-                    <RequireAuth>
-                        <Chatbot userInfo={userInfo} />
-                    </RequireAuth> } />
-                <Route path="/settings" element={
-                    <RequireAuth>
-                        <Settings userInfo={userInfo} />
-                    </RequireAuth> } />
-                <Route path="a/dashboard" element={
-                    <RequireAuth>
-                        <ArtistDashboard userInfo={userInfo} />
-                    </RequireAuth> } />
-                <Route path="a/trends" element={
-                    <RequireAuth>
-                        <TrendingScreen userInfo={userInfo} />
-                    </RequireAuth> } />
                 <Route path='/signup' element={<SignUp setJWT={setJWT} />} />
                 <Route path='/verify/:token' element={<EmailVerification />} />
                 <Route path='/callback' element={<SpotifyCallBack />} />
-				<Route path='/error' element={<Error />} />
+                <Route path='/error' element={<Error />} />
+                <Route element={<RequireAuth><DashboardLayout userInfo={userInfo} handleLogout={handleLogout} /></RequireAuth>}>
+                    <Route path="/profile" element={<Profile userInfo={userInfo} />} />
+                    <Route path="/friends" element={<Friends userInfo={userInfo} />} />
+                    <Route path="/trending" element={<TrendingArtists userInfo={userInfo} />} />
+                    <Route path="/favourites" element={<Favourites userInfo={userInfo} />} />
+                    <Route path="/l/dashboard" element={<ListenerDashboard userInfo={userInfo} />} />
+                    <Route path="/tags" element={<TaggingScreen userInfo={userInfo} />} />
+                    <Route path="/chatbot" element={<Chatbot userInfo={userInfo} />} />
+                    <Route path="/settings" element={<Settings userInfo={userInfo} />} />
+                    <Route path="/a/dashboard" element={<ArtistDashboard userInfo={userInfo} />} />
+                    <Route path="/a/trends" element={<TrendingScreen userInfo={userInfo} />} />
+                </Route>
             </Routes>
         </>
     );
